refactor: motorcycle flow

+++ src/Mottu.Api/UseCases/MotorcycleUseCases/MotorcycleUseCase.cs
@@ -1,6 +1,6 @@
 using Mottu.Api.Domain.Entities;
-using Mottu.Api.Infrastructure.Notifications;
 using Mottu.Api.Infrastructure.Repositories.GenericRepository;
+using Mottu.Api.Infrastructure.Services.Notifications;
 using Mottu.Api.Models;
 
 namespace Mottu.Api.UseCases.MotorcycleUseCases;
@@ -56,17 +56,36 @@ public class MotorcycleUseCase : IMotorcycleUseCase
             _notificationService.Add(new Notification(key, "Placa não pode ser nulo ou vazio"));
         }
 
+        //todo: i can register the same plate, the first in lowercase and the second in uppercase
         if (_repository.Exists(x => x.Plate == request.Plate))
         {
             _notificationService.Add(new Notification(key, $"Moto com a placa {request.Plate} já cadastrada"));
         }
     }
 
-    public IEnumerable<GetMotorcycleResponse> Get(string plate)
+    public IEnumerable<GetMotorcycleResponse> Get(string? plate)
     {
         return _repository
             .GetCollection(string.IsNullOrWhiteSpace(plate) ? null : x => x.Plate == plate)
             .Select(x => new GetMotorcycleResponse(x.Id, x.Year, x.Model, x.Plate));
+    }
+
+    //todo: create unit tests
+    public GetMotorcycleResponse? Get(int id)
+    {
+        if(id <= 0)
+        {
+            return default;
+        }
+
+        var motorcycle = _repository.GetFirst(x => x.Id == id);
+
+        if(motorcycle == null)
+        {
+            return default;
+        }
+
+        return new GetMotorcycleResponse(motorcycle.Id, motorcycle.Year, motorcycle.Model, motorcycle.Plate);
     }
 
     public void Update(int id, PatchMotorcycleRequest request)
@@ -100,9 +119,15 @@ public class MotorcycleUseCase : IMotorcycleUseCase
             _notificationService.Add(new Notification(key, "Id não pode ser menor ou igual a zero"));
         }
 
-        if (string.IsNullOrWhiteSpace(request.Plate))
+        if(string.IsNullOrWhiteSpace(request.Plate))
         {
             _notificationService.Add(new Notification(key, "Placa não pode ser nulo ou vazio"));
+        }
+
+        //todo: i can update the same plate, the first in lowercase and the second in uppercase
+        if (_repository.Exists(x => x.Plate == request.Plate))
+        {
+            _notificationService.Add(new Notification(key, $"Moto com a placa {request.Plate} já cadastrada"));
         }
     }
 
@@ -115,6 +140,12 @@ public class MotorcycleUseCase : IMotorcycleUseCase
         }
 
         var motorcycle = _repository.GetFirst(x => x.Id == id);
+
+        if(motorcycle == null)
+        {
+            _notificationService.Add(new Notification("", "Moto não encontrada"));
+            return;
+        }
 
         //todo: verify if dont exists lease
         _repository.Delete(motorcycle);
